requirements.txt
azure-core==1.1.1
azure-cosmos==4.1.0
bidict==0.21.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.8
influxdb-client==1.24.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.15.5
python-engineio==4.0.1
python-socketio==5.1.0
requests==2.22.0
six==1.15.0
urllib3==1.25.7
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0